fixed file naming conventions, completed carsousel tests, in App.js, change "student' to 'teacher' to render button
