+++ src/components/AddToClass/AddToClass.css
@@ -1,0 +1,7 @@
+.btn-add-to-class {
+  background-color: #f3b89c !important;
+  border: none;
+  margin: 10px;
+  color: black !important;
+  border: none !important;
+}
