+++ src/components/MainLayout/Header.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import Header from '../components/MainLayout/header';
+import Header from './header';
 
 describe('Header component', () => {
   test('renders with teacher role', async () => {
