+++ src/components/SingleMoveNav/SingleMoveNav.css
@@ -1,0 +1,5 @@
+.all-moves-and-classes-btn {
+  color: #117b62 !important;
+  text-decoration: none !important;
+  border: none;
+}
