+++ src/pages/SingleMove/SingleMovePage.css
@@ -1,0 +1,5 @@
+@media (min-width: 1200px) {
+  .col-xl-6 {
+    width: 41% !important;
+  }
+}
