--- src/pages/SingleMove/singleMove.css
@@ -1,19 +1,0 @@
-.btn-add-to-class {
-  background-color: #f3b89c !important;
-  border: none;
-  margin: 10px;
-  color: black !important;
-  border: none !important;
-}
-
-.all-moves-and-classes-btn {
-  color: #117b62 !important;
-  text-decoration: none !important;
-  border: none;
-}
-
-@media (min-width: 1200px) {
-  .col-xl-6 {
-    width: 41% !important;
-  }
-}
